Create a util folder and organize the dist folder.
